requirements.txt
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
psycopg2-binary==2.9.3
py-algorand-sdk==1.16.0
pycodestyle==2.9.0
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
